Leverage cargo's new build command support
This commit adds a build script and a new package to curl-rust itself. The new
dependency, `curl-sys`, is now the house for all FFI-related definitions (it's
basically a header file), and it also contains a build script to detect and find
libcurl. The build script will start to use `pkg-config` to find libcurl, and
failing that will build libcurl from source. This primarily allows this package
to compile on Windows as well as OSX/Unix.