Add a new `Handler` trait and `Easy2` type
One longstanding technical burden with the `Easy` type is that the owned version
requires all callbacks are `Send + 'static`. This particularly wreaks havoc in
the async ecosystem where `Transfer` is unusable. The most onerous bound here
is the `Send` one, typically futures and async things are `'static` already. The
`Send`, however, is typically unnecessary as a handle is rarely sent across
threads in an async scenario.

This commit introduces a new method of handling callbacks that are configured
with libcurl. Instead of specifying callback functions instead a trait
implementation is used. This trait, `Handler`, represents all callbacks that
libcurl can invoke. The trait has default implementations of all callbacks as
"noops" in the same manner that libcurl itself does as well. This means that
only necessary functions need to be implemented.

A new type was added as well, `Easy2<H>`. This easy handle owns an underlying
`H: Handler` and allows easy handles that are neither `'static` nor `Send`. This
way async situations can use the `Handler` trait to avoid these bounds if
necessary.

The `Easy2` API is an exact mirror of the old `Easy` API, and all of the
functionality in the `Easy` module has now been implemented in terms of the
`Easy2` type.

Closes #139