Pass null as form contents if length is empty
Due to a curl oddity if the length is 0 it will call strlen on the
string.  Because there are no trailing nullbytes in rust this will
effectively cause a crash for empty form data.  The workaround is
to ensure that a single null byte is passed for empty form data.

This fixes #176