Add experimental support for rustls (#374)
* Add experimental support for rustls

This adds experimental support for using rustls as a TLS backend for curl. This is made possible by the upstream rustls support recently merged into master. As such, this branch also updates the bundled curl version to the bleeding edge.

We need to use a fork of the crustls wrapper layer for now in order to depend on it in an ideal way. I will work on upstreaming these changes before this can be merged.

There's still some things that will need to be solved before we make this available for everyone, but I am already able to run a sample program with HTTPS successfully under rustls, which is very promising!

* Formatting

* Formatting

* Add feature to readme

* Use upstream version of crustls

* Update to latest main

* Update to bleeding edge

* Add rustls test to CI

* Bump to 0.8.2, disable bundled log capture

* Use rustls-ffi version from Crates.io