curl-sys: remove Haiku cross-compilation workaround
Previously, there were issues cross-compiling the curl library from
source on the Haiku platform. That meant that a workaround was added to
`build.rs` to always link to the shared library in that case. When
testing for the current build, this workaround is no longer necessary,
as cross-compiling the curl-sys crate works as expected.

This removes the workaround for this scenario, as it can now use the
standard behavior.