Fix upload_lots for 8.4.0
This test is hanging on Linux with the 8.4.0 update.

My poor understanding is that `perform()` should not be used with
`curl_multi_socket_action` (that is, it should only use `perform` or
`action`, not both). Removing this seems to fix things.

Note that there were some other hang-related changes in the 8.4.0
release. I bisected the hang to https://github.com/curl/curl/pull/11939,
which then uncovered some curl tests that were hanging in
https://github.com/curl/curl/issues/12033. That then spawned off
two more changes to fix the hangs (both of which are also in 8.4.0).
However, this upload_lots was still failing. I think this is likely
due to an incorrect use of the API poorly interacting with the timer
changes in 8.4.0.