Update to 8.10.1 (#573)
* Update to 8.10.1

* Disable rustls windows cross-compile test

Rustls no longer works for cross-compiling due to switching to aws-lc
(from ring). It seems to want to run `.bat` files for some NASM-related
stuff.

We could, in the future, expose `ring` as a rustls backend as a feature
if someone wants to continue using that.

* Don't try rustls on i686-pc-windows-msvc

It requires nasm to be installed, and I don't want to hassle with that
right now.