Squelch deprecated warning in systest
This removes checking of CURLOPT_WRITEINFO and CURLOPT_CLOSEPOLICY. They
are generating noisy deprecation warnings in systest. These were
deprecated back in 8.10.

I don't think we need to be checking the values of these deprecated
options because they are, well, deprecated, and they shouldn't be used.
At least for now, we know that their values are correct as of 8.14.1
(the 9999 sentinel value).