Use LazyLock for openssl probe (#607)
This switches the global static openssl probe to use LazyLock instead of
Once. `static mut` refs are now an error in Rust 2024, and this is I
believe the preferred method of doing something like this.

LazyLock was added in 1.80 (released 2024-07-25), so the minimum version
has to be bumped to that.